// packages/next-sw/src/types.ts
export type StrategyName =
  | "CacheFirst"
  | "CacheOnly"
  | "NetworkFirst"
  | "NetworkOnly";

export interface MatchCallbackOptions {
  request: Request;
  url: URL;
  sameOrigin: boolean;
}

type UrlPattern = RegExp | string | ((args: MatchCallbackOptions) => boolean);

export interface RuntimeCaching {
  /**
   * This match criteria determines whether the configured handler will
   * generate a response for any requests that don't match one of the precached
   * URLs. If multiple `RuntimeCaching` routes are defined, then the first one
   * whose `urlPattern` matches will be the one that responds.
   */
  urlPattern: UrlPattern;
  /**
   * This determines how the runtime route will generate a response.
   * To use one of the built-in strategies, provide its name,
   * such as `'NetworkFirst'`.
   * Alternatively, this can be a callback function that returns a
   * response.
   */
  handler: StrategyName | ((args: MatchCallbackOptions) => Response);
  /**
   * The HTTP method to match against. The default value of `'GET'` is normally
   * sufficient, unless you explicitly need to match `'POST'`, `'PUT'`, or
   * another type of request.
   * @default "GET"
   */
  method?: string;
  options?: {
    /**
     * If provided, this will set the `cacheName` property of the
     * configured in `handler`. Note that if a cacheName is defined
     * multiple times, only the first entry will be picked.
     */
    cacheName?: string;
    /**
     * Not supported yet.
     */
    expiration?: {
      maxEntries?: number;
      maxAgeSeconds?: number;
    };
    /**
     * If provided, this will set time (in seconds) to wait before timing out a
     * request. Note that only `'NetworkFirst'` and `'NetworkOnly'` support
     * `networkTimeoutSeconds`.
     */
    networkTimeoutSeconds?: number;
  };
}